requirements.txt
streamlit
pandas
pillow
plotly
pytesseract
